#2640 - PT CSGD Award Configuration change (#2904)
Updated the part-time assessment bpmn to have the min of
calculatedDataTotalRemainingNeed2 or the federalAwardCSGDAmount -
programYearTotalPartTimeCSGD.


![image](https://github.com/bcgov/SIMS/assets/62901416/3f443184-249d-4fb4-80ae-0ff23f859f1c)

Since federalAwardCSGDAmount takes care of the
dmnPartTimeAwardAllowableLimits.limitAwardCSGDAmount and the
offeringWeeks multiplication on the previous step. It was removed here.

Also the test case 
- eligibility/parttime-assessment-2021-2022-eligibility-CSGD.e2e-spec.ts
- eligibility/parttime-assessment-2022-2023-eligibility-CSGD.e2e-spec.ts
- eligibility/parttime-assessment-2023-2024-eligibility-CSGD.e2e-spec.ts
which are returning null and failing was removed as the eligibility
check is enough for its validity.

Existing tests cases where enhanced to test the federalAeardNetCSGD
amount and the finalFederalAwardNetCSGDAmount.