#2815 - Disbursement File Code Type Mapping (Part 1) (#2890)
### Changed the assessment summary returned from the API 

The samples for the different outputs can be found along the E2E tests.

- Included the federal and provincial loans. Please keep in mind that
there are no provincial loans for part-time.
- If the estimated BC total grants match with the BC total grants
received in the receipt the values of each BC Grants is copied to the
summary from the estimated values.
- To allow the distinction between eligible awards vs what was not
received in the receipt, the `finalAward` now contains possible `null`
properties as below.
- A value like `disbursementReceipt1cspt: null` would indicate that the
award was estimated but there isn't a receipt value.
- If the property `disbursementReceipt1cspt` is not present at all
(`undefined`) it means that the award was not eligible (same behavior
from before). This would allow the UI to have the distinction shown in
the next image. In the image CSGP is displayed displayed as not eligible
on both sides wich means that the award is not defined on both sides. On
the other hand, CSPT was estimated and a receipt value was expected,
once the payload contains `null`.


![image](https://github.com/bcgov/SIMS/assets/61259237/edcde02a-4c24-45a7-b224-2536db816a54)

- Removed the `receipt.id` property from the `finalAward` result since
they were misleading and unnecessary after these changes. The property
was first populated by the federal `receipt id` and right after
overwritten by the provincial `receipt id`.

#### Expected UI for BC grants when the total value match


![image](https://github.com/bcgov/SIMS/assets/61259237/fa6fe8e9-46f9-4eca-9dc0-711daadd7d84)

#### Expected UI for BC grants when the total value do not match


![image](https://github.com/bcgov/SIMS/assets/61259237/aa40eb30-b84e-4f5a-b9e7-8331efaec3ab)

### E2E Tests

- Created the method
`saveFakeDisbursementReceiptsFromDisbursementSchedule` to allow the
federal and provincial receipts to be created based on disbursement
awards.

#### API Scenarios
- Should get the student assessment summary containing loan and all
grants for a part-time application with a single disbursement.
- Should get the student assessment summary containing loan and all
grants for a part-time application with two disbursements.
- Should get the student assessment summary containing federal and
provincial loans and all grants for a full-time application with two
disbursements.
- Should get the student assessment summary containing federal,
provincial loan, all federal grants and no provincial grants values for
a full-time application when the BCSG does not match.

### Unit tests
- DisbursementReceiptService-getLoanAwardCode
- Should return 'CSLF' when fundingType is 'FE' intensity is 'Full
Time'.
- Should return 'BCSL' when fundingType is 'BC' intensity is 'Full
Time'.
- Should return 'CSLP' when fundingType is 'FE' intensity is 'Part
Time'.
- Should return 'null' when fundingType is 'BC' intensity is 'Part
Time'.

### TODO - Upcoming PR
- Map the award codes.
- DB migration to adjust DB comments.
- E2E tests for file integration.