#2685 - Import Provincial Restrictions from SFAS into SIMS [Bug Fix 2] (#2916)
### As a part of this PR, the following two bugs were fixed:
-----------------------------------------------------------------------
- **Bug:** The Resolved Column was not showing the date for Resolved
restrictions and was instead showing a Date for the Active Restrictions.
**Fix:** It has been fixed to show the resolved date for the Resolved
restrictions and not show a date under the Resolved column for the
currently active restrictions.
<img width="1918" alt="image"
src="https://github.com/bcgov/SIMS/assets/7859295/cd74f2bd-f32b-4f91-8858-3d7d45062e93">

---------------------------------------------------------------------
- **Bug:** The `Bulk-insert-sfas-mapped-restrictions.sql` query was not
checking for the scenario when a specific restriction for a student
mapped from SFAS is inactive is SIMS and at the same time, the same
restriction has entries as an active restriction. If this situation was
there, the query was still allowing for this restriction to be inserted.

**Fix:** This query was adjusted to make sure that if a restriction is
found in the `student_restrictions` table that was mapped from the
`sfas_restrictions` table but is inactive, then the query now checks
that there are no other entries of this same restriction for this
particular student that are present in the `student_restrictions` table
that are currently active. If active restriction entries are found, then
this restriction won't be inserted.
<img width="1080" alt="image"
src="https://github.com/bcgov/SIMS/assets/7859295/f80c7720-a502-46eb-af86-5fad111e6aea">