#2639 - Award Configuration PT - CSPT (#2968)
PT award configuration CSPT updated to consider the program year totals
of CSPT already given to the student.


![image](https://github.com/bcgov/SIMS/assets/62901416/4d478842-487b-469b-8c44-5e36aea94b05)

Bug resolved in this PR:

PT award configuration variables 
- federalAwardNetCSGPAmount
- federalAwardNetCSGDAmount
- provincialAwardNetBCAGAmount
- provincialAwardNetSBSDAmount

all became null when they do not find the below values for the student
in the particular year (or if its the first application for a student in
a PY) so it will return as null
- programYearTotalFullTimeSBSD
- programYearTotalPartTimeSBSD
- programYearTotalFullTimeCSGP
- programYearTotalPartTimeCSGP
- programYearTotalPartTimeCSGD
- programYearTotalPartTimeSBSD

So fixed it by checking a null for the above values and return 0 if they
are null.


![image](https://github.com/bcgov/SIMS/assets/62901416/6fa10019-3116-4d48-943c-f78f982b5f0e)

E2E Test cases  updated