#2685 - Import Provincial Restrictions From SFAS to SIMS - Part 2 (e2e tests) (#2965)
### As a part of this PR, the following e2e tests were written:

- Should save only one legacy restriction for the student if one or more
restrictions are imported from SFAS which do not have a mapping to the
SIMS restrictions.
- Should not insert a restriction to SIMS student restrictions if the
same restriction is imported from SFAS but is already active in SIMS,
Should not insert a legacy restriction if a legacy restriction is
already active in SIMS for the student, Should insert a restriction
after mapping it to SIMS restriction if it is either inactive or is not
at all present in the SIMS student restrictions, Should not insert a
restriction if it has a restriction resolution date.
 
### **Screenshot:**

<img width="1374" alt="image"
src="https://github.com/bcgov/SIMS/assets/7859295/8f8cf01a-cbab-4a10-88f3-731b9839077a">