#2580 - Data Inventory report analysis - next step (#2974)
The spreadsheet for the report is in DEVS --> Files --> Reports.

## In this PR:

- Analyzed most of the columns sources/joins; Some of the outstanding
columns still require some analysis and some require the data to be
saved in the DB as it is calculated by the workflow. Some of the columns
were moved to "Nice to have" and some were removed as per discussions
with the business analysts.
- Changed some names to better describe or fit in 63 characters and
updated the spreadsheet;
- Changed query to select all applications records (Overwritten
included);
- Performance has improved as some subqueries were moved to joins;
- Tried using crosstab to create pivot tables. That had a good
performance on DEV DB but could not use on my local setup as it used
100% of disk for a long time and didn't bring any result for after more
than 5 minutes. That requires more investigation to understand why it
didn't work on our local env.

## Query using crosstab

Here's the version using crosstab:
[crosstab_version.sql.txt](https://github.com/bcgov/SIMS/files/14544045/crosstab_version.sql.txt)

## Some knowledge acquired during the analysis:

### crosstab:
Good example:
https://stackoverflow.com/questions/49051959/postgresql-crosstab-simple-example

That is an example of pivot table generated with crosstab:

```sql
SELECT *
FROM crosstab(
	$$SELECT 
		ds.student_assessment_id,
		vc.value_code,
		coalesce(sum(dv.value_amount), 0) AS value_amount
	FROM 
		(VALUES ('BCAG'),('BCSL'),('BGPD'),('CSGD'),('CSGF'),('CSGP'),('CSGT'),('CSLF'),('CSLP'),('SBSD')) AS vc(value_code)
	CROSS JOIN 
		(SELECT id, student_assessment_id FROM sims.disbursement_schedules) AS ds
	LEFT JOIN 
		sims.disbursement_values dv ON vc.value_code = dv.value_code 
							   AND ds.id = dv.disbursement_schedule_id
	group by ds.student_assessment_id, vc.value_code;$$
) AS ct(student_assessment_id INT, BCAG numeric, BCSL numeric, BGPD numeric, CSGD numeric, CSGF numeric, CSGP numeric, CSGT numeric, CSLF numeric, CSLP numeric, SBSD numeric)
``` 

![image](https://github.com/bcgov/SIMS/assets/78114138/fdee9785-bc47-45a8-b7de-581a612c24c7)


### with

Keeps a query as a temporary table in memory. 

[See more.](https://www.postgresql.org/docs/current/queries-with.html)