#2171 Revise Content of Partner Self-Report Feature (#2996)
- Modified the question "Does your partner have a valid Social Insurance
Number?" to "Is your partner eligible for a B.C. Services Card?"
(Applies to PT and FT applications).
- Modified the help text to "Partners must provide their information to
StudentAid BC by logging in with a B.C. Services Card. If they are
ineligible for a B.C. Services Card (for e.g. if they do not have a
valid SIN), select "No" to this question and provide information on
their behalf." (Applies to PT and FT applications).
- Modified the warning text to "If partners cannot access a B.C.
Services Card, the student will need to report partner information on
their behalf. Please be aware this information will be subject to audit
and review by StudentAid BC and may delay a student's application"
(Applies to PT and FT applications).
- Renamed the form IO API key name "isYourSpouseACanadianCitizen" to
"isYourPartnerAbleToReport" (Applies to PT and FT applications).
 - Corrected content typos "the your".

Screenshot of the help text and the question "Is your partner eligible
for a B.C. Services Card?"

![image](https://github.com/bcgov/SIMS/assets/148148914/13cd5eba-f4a5-4fb9-a69a-ef79ac456fa2)

Screenshot of the warning text and the question "Is your partner
eligible for a B.C. Services Card?"

![image](https://github.com/bcgov/SIMS/assets/148148914/119419bd-aeec-4348-9814-46a0acc72f34)

Screenshot of the partner income page for full-time application
highlighting "your" as corrections to previous typos "the your"

![image](https://github.com/bcgov/SIMS/assets/148148914/8f1a9e0c-f7b5-446c-94f1-f8679c9b1ba5)