#2645 - Program Deactivation - DB migrations (#3004)
Created the new columns related to the education program deactivation.
Used one of the suggested name combinations in the ticket technical
discussion.
- isActive
- isActiveUpdatedBy
- isActiveUpdatedOn

_Minor comment, we did not plan to add a specific note id associated
with the is_active column. The note requested will be created without it
and it was my understanding that it would not be required. Please let me
know if it need to be discussed._

_Minor update in the shared workspace `cSpell.words`._