#2645 - Program Deactivation - API Changes (#3023)
- Created new API endpoints to deactivate the program available for
Institutions and the Ministry. The controller was created to
specifically deactivate the program and the service was created to be
generic to also allow the reactivation in case we need to work in the
requirement in the future.
- Created a new role `institution-deactivate-education-program`.
- Created E2E tests for the new endpoints.
  - Ministry
- Should allow an education program to be deactivated when the program
is active.
- Should return an unprocessable entity HTTP status when the program is
already deactivated.
- Should return a not found HTTP status when the program id does not
exists.
    - Should return a bad request HTTP status when payload is invalid.
  - Institutions
- Should allow an education program to be deactivated when the program
belongs to the institution and the program is active.
- Should return a not found HTTP status when a different institution
accesses the program.
- Should return an unprocessable entity HTTP status when the program is
already deactivated.