#2645 - Program deactivation UI - Part 2 (#3039)
#### Added a new actions menu for the programs available for
Institutions and the ministry.

Institutions will see a label for "Edit program" or "View program" based
in the fact that the program is editable or not.
Ministry will always see the label as "View program".


![image](https://github.com/bcgov/SIMS/assets/61259237/6d75e812-d380-4abe-b800-67b1835b67d1)

When a program is `Inactive` the badges are adjusted and the menu
options are changed to "View program" and the `Deactivate` option is
disabled.


![image](https://github.com/bcgov/SIMS/assets/61259237/655b2e97-9621-44be-8961-30cee860488d)

Badges were also adjusted for the summary view, as shown below.


![image](https://github.com/bcgov/SIMS/assets/61259237/f2421177-d651-4939-be06-d45bc7fa319c)

Program visualization adjusted to display the form is read-only when
Inactive.
_An existing and known error was affecting the form.io to display the
program in read-only. The right visualization is read-only works
depending in the order that the form.io definition and program data are
loaded from the API. To fix it a main container was added to the form.io
definition._


![image](https://github.com/bcgov/SIMS/assets/61259237/97cc0963-f0a6-4c99-8ee3-4702e6a1d899)

- API validation was added to prevent saving an inactive program.
- Created DB migration to allow the SABC code validation only for active
programs. Following similar features, a unique index was used instead of
a DB constraint to allow the `where` condition in DB validation.
- Minor CSS styles were created to allow the bold style to be applied to
the menu action items.
- Created a `danger` global style for Vue following the same color as
`Error`. The goal is to have a better context while setting the color.
- The same modal is shared for Ministry and Institution. The existing
component was already executing client-type checks (they were kept, just
refactored).

**Ministry Modal**

![image](https://github.com/bcgov/SIMS/assets/61259237/20514563-aedd-4389-95c1-02f9ef97deba)

**Institution Modal**

![image](https://github.com/bcgov/SIMS/assets/61259237/b014b96f-70fa-42f0-95f8-aa0c0acdd387)

### Modal Existing Issue
The latest developed modals introduced using the `keepModalOpen` option
have an issue when the API call fails and the API call will not be
executed again if the user retries the operation. This happens because
the promise is not held once the API fails.
To resolve this issue and achieve the same behavior it was introduced
the `beforeResolvePromise` callback that allow the modal to stay open
and the retry operation to work. Please take a look and share some
thoughts.

#### Bulk uploaded constraint adjustment
The bulk uploaded query to search programs was adjusted to include the
`isActive` in the clause allowing the build upload to work also
considering the deactivated programs.