#3018 - PT Lifetime Maximum (Balance) File Import (#3036)
Created a new scheduler StudentLoanBalancesScheduler to be executed once
a day to process the student loan balances file.
Created a table sims.student_loan_balances with columns as below.
    id
    student_id
    csl_balance
    balance_date
    and the other default audit columns.

File expected to be under MSFT-Response folder checked with business and
got confirmation from business it should be in the same folder as the
eCert feedback files and daily disbursement receipts file..
File name is expected to be PEDU.PBC.PT.MBAL.yyyymmdd.001 where the P
stands for the environment.
Validations that are done when the file is read.

    The number of records should match the trailer record count.
    Check for record type. The header is 0 and the footer is 9.
    If some of the above failed, abort the file processing.

Validation that are done in record data (date and balance number)

Validate if the balance is a valid number (check existing similar
validations).
Validate if the date is a valid date (check existing similar
validations).
    If some of the above failed, abort the file processing.

Matching of the student is done by the below criteria (case
insensitive).

    SIN -> student current sin from sin_validations table -> true
    Last Name -> users.last_name
    Birthdate - >students.birth_date
    If no records match, then skip and move to the next record.

If successfully processed, file is deleted. If the process is aborted at
any time, do not delete the file.
If more than one file exists files are processed sequentially ordered by
name.
E2E tests.