#3076 - Application exception bug fix (#3087)
# Fix the correlation id sent on application exception verified message

## Bug
When application exception is verified by ministry (Approved or
Declined), message is sent to workflow with correlation id, so that the
workflow can proceed for further processing.

The workflow expects a message with name
`application-exception-verified` and correlation id `${applicationId}`
but it was receiving a message with correlation id
`${applicationExceptionId}` and hence the message was not accepted and
processed by workflow.

## Fix

Updated the code to send `applicationId` as correlation id for the
message.


![image](https://github.com/bcgov/SIMS/assets/54600590/7ff4439e-c12c-465e-b35f-0dac2a9316df)