#2958 - Part-time NOA Summary Based on e-Cert Values - Part 1 (#3124)
- The dynamic population of the final awards was kept and the changed
code is supposed to follow the same approach already in place, to be
clear.
- Part-time final awards are now retrieved from the disbursement
effective amount (e-Cert effective values).
- The controller service now has a specific method to populate part-time
and a specific method for full-time (that was already in place and was
not changed).
- Date sent added to the result payload for the Ministry and
Institution.
- There was a previous check for COE completion that is now replaced for
the disbursement status defined as `Sent`. which fits Full-time and
Part-time since both need at least the e-Cert generated to have the
final awards populated.

### E2E Tests
- Adapted existing E2E tests.
- Added E2E tests for the Ministry (tests were missing for AEST and they
were added based in the existing ones for student/institution)
- Should get the student assessment award summary containing all federal
and provincial estimated and final awards for a full-time application
with two disbursements when the receipts are available.
- Should get the student assessment award summary containing estimated
and final awards for a part-time application with two disbursements when
e-Cert was sent.
  - Should throw not found exception when assessment is not found.
- Add new E2E test for Students
- Should not generate final award values for a part-time application
when the disbursement has not been sent yet.

### Bug fix
- The disbursement `date_sent` was mapped to the wrong DB date type,
causing the wrong time to persist in the database.

### Upcoming PR
- UI-related changes.