#2872 - PartTime Lifetime maximum check - Part 1 (#3121)
- The latest CSLP balance value is loaded in the Verify assessment
calculation order`.
- E-Cert validation is done as a new step instead of accomodating it in
validateDisbursementPartTimeStep(as mentioned in the Technical part of
the task), as I felt validateDisbursementPartTimeStep was doing more of
the restriction check, so created a new step
validateCSLPLifetimeMaximumCheckStep which handles this check
seperately, only for `part-time` and will block it, if lifetime maximum
CSLP is less than or equal to the disbursed CSLP + latest student loan
balance.

- Check if `calculatedDataTotalRemainingNeed4` is the best reference or
we can execute the logic, for instance in
`finalFederalAwardNetCSLPAmount` calculation.

![image](https://github.com/bcgov/SIMS/assets/62901416/aec92877-f136-4853-b1a7-45a74b13095d)

- Save the `dmnPartTimeAwardAllowableLimits.limitAwardCSLPAmount` to the
`workflow_data` under a new category called `dmnValues` (or similar) to
allow the limit calculation at thee-Cert time.

![image](https://github.com/bcgov/SIMS/assets/62901416/312a7af0-7edd-4578-a1a8-ea710cf46bed)

- Existing failing test cases are fixed for now.

E2E test cases will be done as part of the next PR.