#2689 - Retrieve SFAS/SAIL Award Values Needed for PT PY Maximums (#3120)
- Handle all award values imported from SFAS/SAIL if the award values
are attached to the program year.
- Use `sfas_part_time_applications` and `sfas_applications` tables to
query and fetch the sum of grants.
- Set the start date as the program year start date and the end date as
the first calculated assessment date for the current application.
- Added E2E test scenarios to the `verify-assessment-calculation-order` 
- "Should sum the grants and awards from two past applications with
different offering intensities and SFAS/SAIL applications data when the
applications are for the same student and program year."
- "Should sum the awards from SFAS/SAIL applications data when there is
no past application and the applications are for the same student and
program year."

Screenshot of the assessment values with SFAS part time application
record and without.

![image](https://github.com/bcgov/SIMS/assets/148148914/bdea1890-e260-4c03-b759-ac9275a3b4fd)