Fix Webpack Config and Enable Multiple Debug (#3188)
Over the past weeks, the debug was not working properly due to the
`devtool` configuration (that is the assumption). The configuration we
have right now is the [right one for
production](https://webpack.js.org/configuration/devtool/#devtool) and
to produce the correct stack trace for errors but it was preventing the
debugging from properly working.
A further investigation was not done and the current change is not
supposed to affect production, hence it is intended only to quickly
bring back the debug functionality.

Besides that, new configurations were added to allow the debugging of
the backend applications. Below are the benefits.
- Start the application with the debugger attached.
- Allow application modification keeping the debugger attached.
- Allow the debugging of multiple applications at the same time.

The below image display `api`, `workers`, and `queue-consumers` executed
at the same time with the debugger attached to all of them.


![image](https://github.com/bcgov/SIMS/assets/61259237/1fc8f27c-3b30-413e-becd-d255105ca056)