#2589 - Use a CSV parser library for the report service (#3194)
The backend is currently using the [papaparse
](https://www.npmjs.com/package/papaparse) lib to read CSVs from the
bulk file uploads and the same library has a method to convert objects
to CSV ([unparse](https://www.papaparse.com/docs)).

Used the `Data Inventory Report` as a POC, changing a user first name to
a string with characters that would break the CSV generation.

- Input string: `Testing "quotes" and comma, and more; characters, and
values $45.123,21!`
- Escaped string: `"Testing ""quotes"" and comma, and more; characters,
and values $45.123,21!"`

#### Generated Report

![image](https://github.com/bcgov/SIMS/assets/61259237/4fa78873-de01-448c-b298-5111165e3fa8)

#### CSV escaped string

![image](https://github.com/bcgov/SIMS/assets/61259237/16869fe0-cd7e-40eb-abc6-a6d420b1458a)

#### Same report without using the lib

![image](https://github.com/bcgov/SIMS/assets/61259237/83fe93eb-6c42-4236-9a92-71bd81bc7955)

#### Header is also able to support quotes and other characters


![image](https://github.com/bcgov/SIMS/assets/61259237/968d5460-b0fd-400e-b697-800bc46ffd0a)

Typeorm will produce an object as below. The property names are enclosed
in single or double quotes depending on their content. Please see the
difference between `'Application Number'`, `'First Name with quotes " in
the name and ,coma'`, and `Gender` (this explanation is not part of the
PR and it was added here just for curiosity).
```ts
 {
    'Application ID': 73,
    'Application Number': '5007443876',
    'Application Status': 'Completed',
    'Assessment Date': null,
    'Re-assessment Indicator': 'Original assessment',
    "Marital Status (or 'category')": 'SI',
    'Student Dependency Status': 'dependant',
    'First Name with quotes " in the name and ,coma': 'Maud',
    'Last Name': 'Robel',
    'Date of Birth': 2018-08-01T00:00:00.000Z,
    Gender: 'X',
    SIN: '428314025',
}
```

##### Notes
1. lib was updated to the latest version to take advantage of the
opportunity to use the latest package version.
2. The downloaded file is encoded in UTF-8 as the result of writing it
to the stream, no changes in this behavior since the lib is currently
converting the object to a CSV string that is later written to the
desired output.