Queue-consumers Fail to Start  (#3202)
The PR #3188 introduced an issue where the SQL files expected to be
copied by the `copy-webpack-plugin` during the `webpack` build are no
longer copied.
The current check was failing because the ENV variable was not present
at "build time" and new syntax was also affecting the `plugins`
configuration.

A new env variable (`DEBUG`) was introduced only to allow the `webpack`
to be aware of the `devtools` expected value.