#2366 - ECert Feedback file Part 1 (#3212)
# ECert feedback file changes - Part 1

## New table for the errors that can be received in feedback errors file

- [x] DB Migration to create `sims.ecert_feedback_errors` and insert the
feedback error code and description based on the spread sheet provided
by business.
 Note: The error codes are distinct to an offering intensity (FT | PT).

**FOR REVIEWERS:** Please make use of the attached spreadsheet create by
me based on **_business spread sheet_** to help the code review

[Code_reviewer_helper_guide_for_error_codes.xlsx](https://github.com/bcgov/SIMS/files/15191164/Code_reviewer_helper_guide.xlsx)

## Update the existing table that has disbursements and it's respective
error

- [x] Add a column in `sims.disbursement_feedback_erros` (existing
table) to map the `ecert_feedback_error_id` (foreign key to
`sims.ecert_feedback_errors` ) based on the `error_code` column.
- error_code column will be eventually removed from this table and
`ecert_feedback_error_id` will be the source of error code.
Note: Error_code column is not removed in this PR and it will be removed
in upcoming PR.

- [x] The new column in disbursement_feedback_erros
`ecert_feedback_error_id` is populated with reference from
`sims.disbursement_feedback_erros` based on error_code and once after it
is populated, then ecert_feedback_error_id is made as not nullable

**Note:** Double checked (DEV, TEST and STAGING DB if there is any error
code in the `disbursement_feedback_errors` table which doesn't belong to
`ecert_feedback_errors` to make sure migrations do not fail and there is
no such data.

## E2E Test adjustment for an existing test
- [x] One of the IER E2E test was failing as `ecert_feedback_error_id`
is not nullable and it has been taken care.

## Sonar Cloud duplication issue

There are 257 SQL inserts, where the offering intensity `Full Time` is
repeated 95 times and `Part Time` is repeated 162 times. Sonar cloud is
advising us to use a constant.
 
I haven't added a constant to it as I do not see a great benefit of
using a constant here other than making sonar cloud analysis pass. But I
am open to suggestions and can put effort towards that if team thinks it
is essential.


![image](https://github.com/bcgov/SIMS/assets/54600590/badd8fa9-8bed-4984-a079-ef3d68f4e285)