#3040 - Revise Scholastic Standing Functionality and Content (#3295)
- Adjusted the radio button option for Part-time and Full-time for
`Student did not complete program` (option 3, see images below). Added
hidden field to the already populated
`data.applicationOfferingIntensity`. Using the calculation, it checks if
the label needs to be adjusted. It is defined as Full-time and it
changes to Part-time, if required.
- Accommodate the content the business requires for the individual
options under a panel with generic content for Part-time/Full-time (see
image below).
- DB migration to rename `Change in intensity` to `Student transfer`.
- New business AC added during the sprint to have IER12 code `ST` for
`School transfer` (IER12 docs updated).
- E2E test for the `Change in intensity` was already in place and had
the scholastic type changed to `School transfer`.

## Part-time

![image](https://github.com/bcgov/SIMS/assets/61259237/b53e0266-23ff-4e16-a450-cdc520fd0a37)

## Full-time

![image](https://github.com/bcgov/SIMS/assets/61259237/d22c1ad5-096a-42d4-ac30-6cf14ffe1e2d)

## Common content for both

![image](https://github.com/bcgov/SIMS/assets/61259237/e9699f0e-ae9b-4b48-b361-1ef1aeaa892a)