#3040 - Allow the first disbursement to be not present (#3330)
Reassessments were not displaying a card while the assessment was `In
progress` because there was an expectation that at least the first
disbursement would always be present, which is true only when the
assessment workflow is completed.

The UI was already completely ready expecting the `first disbursement`
not to be present, but the API was throwing an error every time.

### Current UI without the fix


![image](https://github.com/bcgov/SIMS/assets/61259237/3227a998-3cba-45bc-b810-0655a5f07c56)

### UI after the fix


![image](https://github.com/bcgov/SIMS/assets/61259237/08f95cea-8331-4897-9670-a43551b878a3)