#3282 - Detect Assessment Offering Dates Change - Part 3 (e2e tests) (#3344)
### As a part of this PR, the following e2e tests for the workers were
completed:

AssessmentController(e2e)-workflowWrapUp
√ Should populate the previousDateChangedReportedAssessment when the
e-Cert was generated for the 'Original Assessment' and offering dates
changed as a result of the reassessment.
√ Should not populate the previousDateChangedReportedAssessment when the
e-Cert was generated for the 'Original Assessment' and offering dates
didn't change as a result of the reassessment.
√ Should not populate the previousDateChangedReportedAssessment when the
e-Cert is not generated for the 'Original Assessment' and offering dates
changed as a result of the reassessment.
√ Should populate the previousDateChangedReportedAssessment with the
most recent reported reassessment when there are previously reported
reassessments.
√ Should not populate the previousDateChangedReportedAssessment if there
are multiple reassessments where an intermediate reassessment has a
different offering date but the current one changed the offering date
back to the value of the latest reported assessment.
√ Should not populate the previousDateChangedReportedAssessment if there
are multiple reassessments where an intermediate reassessment has a
different offering date but the current one changed the offering dates
back to the value of the last generated ecert when there are no report
based reported assessments.

### **Screenshot for the above created e2e tests:**

<img width="1360" alt="image"
src="https://github.com/bcgov/SIMS/assets/7859295/b791ee5b-c105-4127-94f0-dbe5c4be5ac6">


------------------------------------------------------------------------------------------------------------------------------------------

**Note:** I have updated the `workFlowWrapUp` test file to use
`restoreAllMocks` rather than the `resetAllMocks` because the
`restoreAllMocks` restores the function implementation to the original
(non-mocked) implementation whereas the `resetAllMocks` replaces the
mock implementation with an empty function, returning undefined. The
first behavior is what was needed for each test and hence used
`restoreAllMocks`.

<img width="843" alt="image"
src="https://github.com/bcgov/SIMS/assets/7859295/8db0244e-9063-4101-aa9b-cabcb5233e98">

Reference: https://jestjs.io/docs/mock-function-api#mockfnmockrestore