#3300 - Income Eligibility - PT CSLP/SBSD (#3371)
- [x] Update awardEligibilitySBSD in all PT program years to include
limitAwardSBSDThresholdIncome

![image](https://github.com/bcgov/SIMS/assets/62901416/231aca61-6201-49a9-a163-06cdd69fcfee)

- [x] Create awardEligibilityCSLP in all PT program years and include in
the federalAwardNetCSLPAmount calculation

![image](https://github.com/bcgov/SIMS/assets/62901416/c9313bad-7f14-4d2a-a11b-d30654dc3fec)

- [x] Include the awardEligibilityCSLP in
parttime-configure-disbursement.bpmn workflow when the payload for
disbursement schedule is created for both 17weeks or 18weeks and more.

![image](https://github.com/bcgov/SIMS/assets/62901416/491d07d2-79d5-4107-b01a-98edcba9adb0)

- [x] E2E Test cases are updated and created.

As mentioned by @andrewsignori-aot in the link
https://github.com/bcgov/SIMS/issues/3300#issuecomment-2148361233. The
pages are unstable if there are not award eligible for a student withour
PD/PPD status. If things has to be changed, new ticket will be created.