#1366 - Set FileOriginType for Student Application Files (#3542)
As per the analysis, student files are never deleted, only the
relationship between an application and its files is deleted while the
application is in draft. Once the application is summited the file
relations are no longer deleted.
According to the confirmation from @CarlyCotton and @Joshua-Lakusta,
there are no concerns about keeping the current behavior since the
application only starts to matter when submitted.

### Minor improvement

Student application files were never updated from their `Temporary`
status because the status was added later.
This PR is now updating the file origin when an application is saved.